Show habits retreived from db
