+++ app/DBFunctions.js
@@ -1,4 +1,8 @@
 import database from '@react-native-firebase/database'
+import * as React from 'react';
+import { Avatar, Button, Card, Title, Paragraph, ProgressBar, Colors } from 'react-native-paper';
+import { NavigationContainer, useNavigation } from '@react-navigation/native';
+
 
 export const addItem = (itemId, itemName) => {
     return new Promise(function(resolve,reject){
@@ -22,7 +26,7 @@ export const addItem = (itemId, itemName) => {
     });
 };
 
-export const addHabit = (habitId, habitDesc, date, numPD, category) => {
+export const addHabit = (habitId, habitDesc, startDate, numPD, category, consPts, points, date) => {
     return new Promise(function(resolve,reject){
         let key;
         if (habitId != null) {
@@ -34,9 +38,12 @@ export const addHabit = (habitId, habitDesc, date, numPD, category) => {
         let dataToSave = {
             habitId:key,
             habitDesc: habitDesc,
-            date: date,
+            startDate: startDate,
             numPerD: numPD,
             category: category,
+            consPts: consPts,
+            points: points,
+            date: date,
         };
         database().ref('habits/' + key).update(dataToSave).then((snapshot)=>{
             resolve(snapshot)
@@ -47,10 +54,94 @@ export const addHabit = (habitId, habitDesc, date, numPD, category) => {
     });
 };
 
+export const getHabit = (props) => {
+    const [habits, setHabits] = React.useState([]);
+    const habitRef = database().ref('/habits');
+    const navigation = useNavigation();
+
+    const [boolean, setBool] = React.useState(false);
+    const [counter, setCounter] = React.useState(1);
+    const onLoadingListener = habitRef.on('value', snapshot =>{
+        setHabits([]);
+        snapshot.forEach(function(childSnapshot){
+            setHabits(habits => [...habits, childSnapshot.val()]);
+        });
+    });
+
+    const show = () => habits.map((element) => {
+        habitRef.off('value', onLoadingListener);
+        return(
+            <Card key={element.habitId}>
+            <Title>{element.habitId}</Title>
+            <Paragraph>
+              Description
+              start date
+              time
+            </Paragraph>
+            <Card.Actions>
+              <Button onPress={() => {
+                props.setPoints(props.points + 10);
+                props.setCPts(props.consPts + 10);
+                reward(props.consPts);
+                setCounter(counter + 1);
+  
+                // number per day retrieved from db
+                const testCounter = 2;
+                const currentDate = new Date();
+                
+                //date retrieved from db
+                const retrievedDate   = new Date("2021-02-07");
+  
+                const diff = (currentDate - retrievedDate) / (1000 * 60 * 60 * 24);
+  
+                // checks if the current date is smaller the the retrieved date 
+                // checks if the user pressed the button more than he should be pressing it
+                if(diff <= 1 || counter >= testCounter){
+                    setBool(true);
+                    setCounter(0);
+                }
+                if(diff > 1 && counter < testCounter){
+                  setBool(false);
+                }
+                }}
+                disabled={boolean}
+                >Done</Button>
+              <Button onPress={() => {
+                if (props.points != 0){
+                  props.setPoints(props.points - 5);
+                }
+                props.setCPts(0);
+                setBool(false);
+              }}>Fail</Button>
+              <Button onPress={() => {
+                
+                console.log(getHabit());
+            navigation.navigate('Tasks', {
+              screen: 'EditHabit'
+            });
+          }}>Edit</Button>
+            </Card.Actions>
+          </Card>
+        );
+      });
+
+      return show();
+    
+    // return () => {
+    //     habitRef.off('value', onLoadingListener);
+    // }
+
+}
+
+export const listHabits = (habits) =>{
+    
+
+}
+
 export const deleteItem = (itemId, deleteConfirm) => {
 
         database().ref('items/' + itemId).remove().then(() => {
-        }).catch((erro) =>{
+        }).catch((err) =>{
             console.log(err)
         });
